checkpoint: add translate module; test still need to be reworked
