+++ basin3d/core/models.py
@@ -30,33 +30,34 @@ from typing import List, Optional, Union
 
 from basin3d.core import monitor
 from basin3d.core.schema.enum import FeatureTypeEnum, FEATURE_SHAPE_TYPES, MAPPING_DELIMITER, NO_MAPPING_TEXT
+from basin3d.core.translate import get_datasource_mapped_attribute, translate_attributes
 from basin3d.core.types import SpatialSamplingShapes
 
 logger = monitor.get_logger(__name__)
 
 
-def translate_attributes(plugin_access, mapped_attrs, **kwargs):
-    """Helper method to translate datasource vocabularies to BASIN-3D vocabularies via MappedAttributes"""
-
-    # copy the kwargs be able to loop thru the original while modifying the actual for compound mappings
-    kwargs_orig = kwargs.copy()
-
-    for attr in mapped_attrs:
-        if attr in kwargs_orig:
-            datasource_vocab = kwargs[attr]
-            attr_mapping = plugin_access.get_datasource_mapped_attribute(attr_type=attr.upper(), attr_vocab=datasource_vocab)
-            kwargs[attr] = attr_mapping
-
-            # If the attr is part of a compound mapping and the compound attr is not part of the kwargs, set it.
-            cm_attrs = plugin_access.get_compound_mapping_attributes(attr)
-            if cm_attrs:
-                for cm_attr in cm_attrs:
-                    if cm_attr.lower() not in kwargs:
-                        cm_attr_mapping = plugin_access.get_datasource_mapped_attribute(attr_type=cm_attr.upper(),
-                                                                                        attr_vocab=datasource_vocab)
-                        kwargs[cm_attr.lower()] = cm_attr_mapping
-
-    return kwargs
+# def translate_attributes(plugin_access, mapped_attrs, **kwargs):
+#     """Helper method to translate datasource vocabularies to BASIN-3D vocabularies via MappedAttributes"""
+#
+#     # copy the kwargs be able to loop thru the original while modifying the actual for compound mappings
+#     kwargs_orig = kwargs.copy()
+#
+#     for attr in mapped_attrs:
+#         if attr in kwargs_orig:
+#             datasource_vocab = kwargs[attr]
+#             attr_mapping = plugin_access.get_datasource_mapped_attribute(attr_type=attr.upper(), attr_vocab=datasource_vocab)
+#             kwargs[attr] = attr_mapping
+#
+#             # If the attr is part of a compound mapping and the compound attr is not part of the kwargs, set it.
+#             cm_attrs = plugin_access.get_compound_mapping_attributes(attr)
+#             if cm_attrs:
+#                 for cm_attr in cm_attrs:
+#                     if cm_attr.lower() not in kwargs:
+#                         cm_attr_mapping = plugin_access.get_datasource_mapped_attribute(attr_type=cm_attr.upper(),
+#                                                                                         attr_vocab=datasource_vocab)
+#                         kwargs[cm_attr.lower()] = cm_attr_mapping
+#
+#     return kwargs
 
 
 class JSONSerializable:
@@ -1432,7 +1433,7 @@ class ResultListTVP(Base):
         # translate quality
         # ToDo: figure out how to handle quality v result_quality discrepancy
         if 'quality' in kwargs and kwargs['quality']:
-            kwargs['quality'] = plugin_access.get_datasource_mapped_attribute(attr_type='RESULT_QUALITY', attr_vocab=kwargs['quality'])
+            kwargs['quality'] = get_datasource_mapped_attribute(plugin_access, attr_type='RESULT_QUALITY', datasource_vocab=kwargs['quality'])
 
         # Initialize after the attributes have been set
         super().__init__(plugin_access, **kwargs)
@@ -1498,7 +1499,7 @@ class ResultPointFloat(Base):
         # translate quality
         # ToDo: figure out how to handle quality v result_quality discrepancy
         if 'quality' in kwargs and kwargs['quality']:
-            kwargs['quality'] = plugin_access.get_datasource_mapped_attribute(attr_type='RESULT_QUALITY', attr_vocab=kwargs['quality'])
+            kwargs['quality'] = get_datasource_mapped_attribute(plugin_access, attr_type='RESULT_QUALITY', datasource_vocab=kwargs['quality'])
 
         # Initialize after the attributes have been set
         super().__init__(None, **kwargs)
